+++ tests/Unit/Services/IM/IMServiceBuilderTest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Bitrix24\SDK\Tests\Unit\Services\IM;
 
+use Bitrix24\SDK\Services\IM\Counters\Service\Counters;
 use Bitrix24\SDK\Services\IM\IMServiceBuilder;
 use Bitrix24\SDK\Services\IM\Placements\Placements;
 use Bitrix24\SDK\Services\ServiceBuilder;
@@ -41,6 +42,12 @@ class IMServiceBuilderTest extends TestCase
     public function testGetMessageService(): void
     {
         $this::assertSame($this->serviceBuilder->message(), $this->serviceBuilder->message());
+    }
+
+    public function testGetCountersService(): void
+    {
+        $this->assertInstanceOf(Counters::class, $this->serviceBuilder->counters());
+        $this->assertSame($this->serviceBuilder->counters(), $this->serviceBuilder->counters());
     }
 
     public function testGetPlacementsService(): void